refactor(platform): ensure_session as one explicit entry, not buried in validate_env
Move ensure_session (+ _login/_auth_url) from benchmax.cli to a new
benchmax.platform.login module and export it from benchmax.platform, so scripts
import it cleanly and the layering is right (cli depends on platform, not the
reverse). Drop the auto-login call I'd put inside validate_env: interactive
login is a single explicit up-front call (the generated script / a human calls
ensure_session once), NOT something library/data-prep functions trigger — they
resolve via the per-request seam and fail loudly. validate_env keeps its
remote-on-local=False gating; it just no longer logs you in.
